feat(lib): rewrite clickOutside directive in TS
Rewrites the directive defined in `src/directives/clickOutside.js` in
TypeScript.

Replaces `navigator.msMaxTouchPoints` with `navigator.maxTouchPoints`.
`msMaxTouchPoints` is for IE10 which we no longer support.

The parameters of `toggleEventListeners` are no longer optional because
the function is internal and every call in the file specifies all the
parameters. The `action` parameter may take only "add" or "remove" so
that a type error at `document[`${action}EventListener`]` is suppressed.

Prepends the `!isFunction` predicate to guarantee `bindingValue` is an
option object:
- `!isFunction && bindingValue.middleware`
- `!isFunction && bindingValue.events`

`instance` should be removed from `instances` on `unmounted`, though,
I decided not to fix it in this commit because we decided to change the
logics as little as possible during the migration process.

Removes `instances` from the exported directive definition object,
because it has nothing to do with the directive definition.

Adds some utility types to ease the type definition.