fix(lib): editable Clockpicker did not work
Fixes the issue that `Clockpicker` with the `editable` prop `true` did
not work. Makes the underlying `Input` lazy if it is editable, and
traps "update:model-value" events instead of "change" events to update
the time. According to my observation, the "change" event handler of
`Clockpicker` is called before the handler of `Input` and cancels
user's input.